+++ js/data.js
@@ -1,5 +1,21 @@
-import {getPost, COUNT_POSTS} from '.data/images.js';
+const MIN_INDEX_AVATAR = () => 1;
+const MAX_INDEX_AVATAR = () => 6;
+const COUNT_POSTS = () => 25;
+const MIN_COUNT_LIKES = () => 15;
+const MAX_COUNT_LIKES = () => 200;
+const MIN_COUNT_COMMENTS = () => 0;
+const MAX_COUNT_COMMENTS = () => 30;
 
-Array.from({length: COUNT_POSTS}, getPost);
+const MESSAGES = () => ['Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
 
+const NAMES = () => [
+  'Иван', 'Екатерина', 'Александр', 'Ольга', 'Дмитрий', 'Анна', 'Михаил', 'Татьяна'
+];
 
+export {MIN_INDEX_AVATAR, MAX_INDEX_AVATAR, COUNT_POSTS, MIN_COUNT_LIKES, MAX_COUNT_LIKES, MIN_COUNT_COMMENTS, MAX_COUNT_COMMENTS, MESSAGES, NAMES };
