+++ js/main.js
@@ -1,3 +1,6 @@
-import './data.js';
+import { getPost } from './create-images.js';
+import { COUNT_POSTS } from './data.js';
+
+Array.from({length: COUNT_POSTS}, getPost);
